requirements.txt
# Core Frameworks (Pinned for 2026 Stability)
langchain==1.2.12
langchain-core==1.2.18
langchain-community==0.4.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
pydantic==2.12.5

# Web & WhatsApp
fastapi==0.110.0
uvicorn[standard]==0.27.1
requests==2.31.0
python-multipart==0.0.9
python-dotenv==1.0.1
pywa==3.9.0

# LLM Providers (Updated for Gemini & OpenAI)
langchain-google-genai>=4.2.1
google-genai>=1.60.0
openai>=1.50.0